Update http_proxy.go to make session cookie "domain" blank
Previously this used the root domain but this broke things when you tried to use something on this list: https://github.com/publicsuffix/list, like "cloudfront.net" because browsers will ignore it and the cookie will not appear in future requests. With a blank domain, the browser assumes it to be the full hostname (i.e. abcdefg.cloudfront.net). This could cause issues depending on the phishlet but at least works for M365 as all the requests related to authentication only use the first subdomain specified in the phishlet.